added factory for kafka
